refactor(account): use new attribute name and type
