+++ src/account/mapper/managed-project.mapper.ts
@@ -15,7 +15,7 @@ export class ManagedProjectMapper {
             oneLiner: project.oneLiner,
             category: project.category,
             accessAddresses: project.accessAddresses,
-            socialMedia: project.socialMedia ?? {},
+            mediaHandles: project.mediaHandles ?? {},
             logo: project.logo
                 ? {
                       id: project.logo.id,
